+++ junopy/services/notifications.py
@@ -9,6 +9,9 @@ def EventTypes():
     return ListType(EventType).add([EventType(**item) for item in data['_embedded']['eventTypes']]) if '_embedded' in data else ListType(EventType)
 
 
-def Webhooks():
-    data = Get("/notifications/webhooks")
+def Webhooks(resourceToken=None):
+    aditional_header = None
+    if not resourceToken is None:
+        aditional_header = {'resourceToken':resourceToken}
+    data = Get("/notifications/webhooks", aditional_header=aditional_header)
     return ListType(Webhook).add([Webhook(**item) for item in data['_embedded']['webhooks']]) if '_embedded' in data else ListType(Webhook)