+++ junopy/services/transfers.py
@@ -3,15 +3,15 @@ from junopy.utils.juno import *
 from junopy import Transfer, BankAccount
 
 
-def Default(amount: float):
+def Default(amount: float, digitalAccountId=None, resourceToken=None):
     __transfer = Transfer(**{
         "type": "DEFAULT_BANK_ACCOUNT",
         "amount": amount
     })
-    return __transfer.Create()
+    return __transfer.Create(resourceToken=resourceToken)
 
 
-def P2P(name: str, document: str, amount: float, accountNumber: str):
+def P2P(name: str, document: str, amount: float, accountNumber: str, resourceToken=None):
     __transfer = Transfer(**{
         "type": "P2P",
         "name": name,
@@ -19,10 +19,10 @@ def P2P(name: str, document: str, amount: float, accountNumber: str):
         "amount": amount,
         "bankAccount": {"accountNumber":  accountNumber}
     })
-    return __transfer.Create()
+    return __transfer.Create(resourceToken=resourceToken)
 
 
-def Bank(name: str, document: str, amount: float, bank: BankAccount):
+def Bank(name: str, document: str, amount: float, bank: BankAccount, resourceToken=None):
     __transfer = Transfer(**{
         "type": "BANK_ACCOUNT",
         "name": name,
@@ -30,10 +30,10 @@ def Bank(name: str, document: str, amount: float, bank: BankAccount):
         "amount": amount,
         "bankAccount": bank
     })
-    return __transfer.Create()
+    return __transfer.Create(resourceToken=resourceToken)
 
 
-def Pix(name: str, document: str, amount: float, bank: BankAccount):
+def Pix(name: str, document: str, amount: float, bank: BankAccount, resourceToken=None):
     __transfer = Transfer(**{
         "type": "PIX",
         "name": name,
@@ -41,4 +41,4 @@ def Pix(name: str, document: str, amount: float, bank: BankAccount):
         "amount": amount,
         "bankAccount": bank
     })
-    return __transfer.Create()
+    return __transfer.Create(resourceToken=resourceToken)